ciandt_it-nextgen_time28-c01df5cce7b2/python_backend/feedbackresponseview/pyproject.toml
[project]
name = "python-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Jhonatan Teixeira",email = "jhonatant@ciandt.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[all] (>=0.115.11,<0.116.0)",
    "peewee (>=3.17.9,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"

